Splitting out SeriesTest to make things clearer.
